refactor: collectionViewDelegate 분리를 위해 코드 수정 photoSelectButton 을 EnrollModifyViewController 로 이동 EnrollModifyCollectionViewDelegate 타입 생성
