+++ OpenMarket/OpenMarket/Controller/EnrollModifyViewController.swift
@@ -12,22 +12,39 @@ class EnrollModifyViewController: UIViewController {
     @IBOutlet weak var postPatchButton: UIBarButtonItem!
     
     private let enrollModifyCollectionViewDataSource = EnrollModifyCollectionViewDataSource()
+    private let enrollModifyCollectionViewDelegate = EnrollModifyCollectionViewDelegate()
     private let delegate = UIApplication.shared.delegate as? AppDelegate
     private var selectIndexPathDictionary: [IndexPath: Bool] = [:]
     private let mainTitle = "상품"
     var topItemTitle: String = ""
+    private let photoTotalNumber = 5
+    private var photoSelectNumber = 0
+    private let cameraImage = UIImage(systemName: "camera")
+    private let photoSelectButton: UIButton = {
+        let button = UIButton()
+        button.contentMode = .scaleAspectFit
+        return button
+    }()
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        photoSelectButton.setTitle(
+                   "\(photoSelectNumber)/\(photoTotalNumber)", for: .normal)
+        photoSelectButton.setImage(cameraImage, for: .normal)
+        
         collectionView.dataSource = enrollModifyCollectionViewDataSource
-        collectionView.delegate = self
+        collectionView.delegate = enrollModifyCollectionViewDelegate
         self.title = mainTitle + topItemTitle
         postPatchButton.title = topItemTitle
         collectionView.register(EnrollModifyPhotoSeclectCell.self, forCellWithReuseIdentifier: EnrollModifyPhotoSeclectCell.identifier)
         collectionView.register(EnrollModifyPhotoCell.self, forCellWithReuseIdentifier: EnrollModifyPhotoCell.identifier)
         collectionView.register(EnrollModifyListCell.self, forCellWithReuseIdentifier: EnrollModifyListCell.identifier)
         collectionView.collectionViewLayout = enrollModifyCollectionViewDataSource.createCompositionalLayout()
+        enrollModifyCollectionViewDataSource.butteeen.append(photoSelectButton)
+        
+        photoSelectButton.addTarget(self, action: #selector(movePhotoAlbum(_:)), for: .touchUpInside)
+        configureEnrollModifyDataSource()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -41,26 +58,21 @@ class EnrollModifyViewController: UIViewController {
         
         delegate?.changeOrientation = true
     }
-}
-
-extension EnrollModifyViewController: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-//        guard let photoCell = collectionView.cellForItem(at: indexPath) as? EnrollModifyPhotoCell else { return }
-        if indexPath.item == .zero && indexPath.section == .zero {
-            guard let convertPhotoAlbumViewController = storyboard?.instantiateViewController(identifier: PhotoAlbumViewController.identifier) as? PhotoAlbumViewController else {
-                return
-            }
-            convertPhotoAlbumViewController.selected = { (image: [UIImage]) in
-                self.enrollModifyCollectionViewDataSource.photoAlbumImages += image
-                collectionView.reloadData()
-            }
-            navigationController?.pushViewController(convertPhotoAlbumViewController, animated: true)
-        } else {
-            guard let photoIndexPaths = collectionView.indexPathsForSelectedItems else { return }
-            for indexPath in photoIndexPaths {
-                enrollModifyCollectionViewDataSource.photoAlbumImages.remove(at: indexPath.item - 1)
-            }
-            collectionView.deleteItems(at: photoIndexPaths)
+    
+    func configureEnrollModifyDataSource() {
+        enrollModifyCollectionViewDelegate.updateEnrollModifyCollectionViewDataSource(enrollModifyCollectionViewDataSource: enrollModifyCollectionViewDataSource)
+    }
+    
+    @objc func movePhotoAlbum(_ sender: UIButton) {
+        guard let convertPhotoAlbumViewController = storyboard?.instantiateViewController(identifier: PhotoAlbumViewController.identifier) as? PhotoAlbumViewController else {
+            return
         }
+        
+        convertPhotoAlbumViewController.selected = { (image: [UIImage]) in
+            self.enrollModifyCollectionViewDataSource.photoAlbumImages += image
+            self.collectionView.reloadData()
+        }
+        navigationController?.pushViewController(convertPhotoAlbumViewController, animated: true)
     }
 }
+
