+++ OpenMarket/OpenMarket/View/EnrollModifyCell/EnrollModifyPhotoSelectCell.swift
@@ -9,15 +9,6 @@ import UIKit
 
 class EnrollModifyPhotoSeclectCell: UICollectionViewCell {
     static let identifier = String(describing: EnrollModifyPhotoSeclectCell.self)
-    private let photoTotalNumber = 5
-    private var photoSelectNumber = 0
-    private let cameraImage = UIImage(systemName: "camera")
-    private let photoSelectButton: UIButton = {
-        let button = UIButton()
-        button.isUserInteractionEnabled = false
-        button.contentMode = .scaleAspectFit
-        return button
-    }()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -29,12 +20,12 @@ class EnrollModifyPhotoSeclectCell: UICollectionViewCell {
         self.layer.cornerRadius = 10
         self.layer.borderWidth = 1
         self.layer.borderColor = UIColor.black.cgColor
-        photoSelectButton.setTitle(
-            "\(photoSelectNumber)/\(photoTotalNumber)", for: .normal)
+        
+    }
+    
+    func configure(photoSelectButton: UIButton) {
         photoSelectButton.setTitleColor(.black, for: .normal)
         photoSelectButton.tintColor = .black
-        photoSelectButton.setImage(cameraImage, for: .normal)
-        photoSelectButton.isUserInteractionEnabled = false
         contentView.addSubview(photoSelectButton)
         photoSelectButton.frame = CGRect(x: 0, y: 0,
                                          width: contentView.frame.width,
